Clean up module Use dedicated class for global configuration Remove PaymentMethod injections Create abstract payplug payment method
