+++ Controller/Adminhtml/Order/UpdatePayment.php
@@ -6,21 +6,21 @@ use Magento\Backend\App\Action;
 use Magento\Sales\Api\OrderManagementInterface;
 use Magento\Sales\Api\OrderRepositoryInterface;
 use Payplug\Exception\PayplugException;
+use Payplug\Payments\Helper\Data;
 use Payplug\Payments\Logger\Logger;
-use Payplug\Payments\Model\PaymentMethod;
 use Psr\Log\LoggerInterface;
 
 class UpdatePayment extends \Magento\Sales\Controller\Adminhtml\Order
 {
     /**
-     * @var PaymentMethod
-     */
-    protected $paymentMethod;
-
-    /**
      * @var Logger
      */
     protected $payplugLogger;
+
+    /**
+     * @var Data
+     */
+    protected $payplugHelper;
 
     /**
      * @param Action\Context                                   $context
@@ -34,8 +34,8 @@ class UpdatePayment extends \Magento\Sales\Controller\Adminhtml\Order
      * @param OrderManagementInterface                         $orderManagement
      * @param OrderRepositoryInterface                         $orderRepository
      * @param LoggerInterface                                  $logger
-     * @param PaymentMethod                                    $paymentMethod
      * @param Logger                                           $payplugLogger
+     * @param Data                                             $payplugHelper
      */
     public function __construct(
         Action\Context $context,
@@ -49,11 +49,11 @@ class UpdatePayment extends \Magento\Sales\Controller\Adminhtml\Order
         OrderManagementInterface $orderManagement,
         OrderRepositoryInterface $orderRepository,
         LoggerInterface $logger,
-        PaymentMethod $paymentMethod,
-        Logger $payplugLogger
+        Logger $payplugLogger,
+        Data $payplugHelper
     ) {
-        $this->paymentMethod = $paymentMethod;
         $this->payplugLogger = $payplugLogger;
+        $this->payplugHelper = $payplugHelper;
         parent::__construct(
             $context,
             $coreRegistry,
@@ -75,15 +75,14 @@ class UpdatePayment extends \Magento\Sales\Controller\Adminhtml\Order
     public function execute()
     {
         if ($order = $this->_initOrder()) {
-
-            if (!$this->paymentMethod->canUpdatePayment($order)) {
+            if (!$this->payplugHelper->canUpdatePayment($order)) {
                 $this->messageManager->addErrorMessage(__('The payment cannot be updated for this order.'));
 
                 return $this->_redirect('sales/order/view', ['order_id' => $order->getId()]);
             }
 
             try {
-                $this->paymentMethod->updatePayment($order);
+                $order->getPayment()->getMethodInstance()->updatePayment($order);
                 $this->messageManager->addSuccessMessage(__('Order payment was successfully updated.'));
             } catch (PayplugException $e) {
                 $this->payplugLogger->error($e->__toString());
