+++ Controller/Payment/AbstractPayment.php
@@ -4,7 +4,6 @@ namespace Payplug\Payments\Controller\Payment;
 
 use Payplug\Payments\Helper\Data;
 use Payplug\Payments\Logger\Logger;
-use Payplug\Payments\Model\PaymentMethod;
 
 abstract class AbstractPayment extends \Magento\Framework\App\Action\Action
 {
@@ -17,11 +16,6 @@ abstract class AbstractPayment extends \Magento\Framework\App\Action\Action
      * @var \Magento\Sales\Model\OrderFactory
      */
     protected $salesOrderFactory;
-
-    /**
-     * @var PaymentMethod
-     */
-    protected $paymentMethod;
 
     /**
      * @var Logger
@@ -37,7 +31,6 @@ abstract class AbstractPayment extends \Magento\Framework\App\Action\Action
      * @param \Magento\Framework\App\Action\Context $context
      * @param \Magento\Checkout\Model\Session       $checkoutSession
      * @param \Magento\Sales\Model\OrderFactory     $salesOrderFactory
-     * @param PaymentMethod                         $paymentMethod
      * @param Logger                                $logger
      * @param Data                                  $payplugHelper
      */
@@ -45,14 +38,12 @@ abstract class AbstractPayment extends \Magento\Framework\App\Action\Action
         \Magento\Framework\App\Action\Context $context,
         \Magento\Checkout\Model\Session $checkoutSession,
         \Magento\Sales\Model\OrderFactory $salesOrderFactory,
-        PaymentMethod $paymentMethod,
         Logger $logger,
         Data $payplugHelper
     ) {
         parent::__construct($context);
         $this->checkoutSession = $checkoutSession;
         $this->salesOrderFactory = $salesOrderFactory;
-        $this->paymentMethod = $paymentMethod;
         $this->logger = $logger;
         $this->payplugHelper = $payplugHelper;
     }
