+++ Controller/Payment/Redirect.php
@@ -25,8 +25,10 @@ class Redirect extends AbstractPayment
                 return;
             }
 
-            $payment = $this->paymentMethod->createPayplugTransaction($order);
-            $this->paymentMethod->setOrderPendingPayment($order);
+            $paymentMethod = $order->getPayment()->getMethodInstance();
+
+            $payment = $paymentMethod->createPayplugTransaction($order);
+            $paymentMethod->setOrderPendingPayment($order);
             $url = $payment->hosted_payment->payment_url;
 
             return $this->resultRedirectFactory->create()->setUrl($url);
