+++ Model/Config/Source/PaymentPage.php
@@ -2,7 +2,7 @@
 
 namespace Payplug\Payments\Model\Config\Source;
 
-use Payplug\Payments\Model\PaymentMethod;
+use Payplug\Payments\Model\Payment\AbstractPaymentMethod;
 
 class PaymentPage implements \Magento\Framework\Option\ArrayInterface
 {
@@ -13,11 +13,11 @@ class PaymentPage implements \Magento\Framework\Option\ArrayInterface
     {
         $options = [
             [
-                'value' => PaymentMethod::PAYMENT_PAGE_REDIRECT,
+                'value' => AbstractPaymentMethod::PAYMENT_PAGE_REDIRECT,
                 'label' => __('Redirect'),
             ],
             [
-                'value' => PaymentMethod::PAYMENT_PAGE_EMBEDDED,
+                'value' => AbstractPaymentMethod::PAYMENT_PAGE_EMBEDDED,
                 'label' => __('Embedded'),
             ],
         ];
