+++ etc/config.xml
@@ -1,24 +1,27 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Store:etc/config.xsd">
     <default>
         <payment>
-            <payplug_payments>
+            <payplug_payments_standard>
                 <active>0</active>
                 <title>Credit card checkout</title>
-                <environmentmode>test</environmentmode>
-                <payment_page>redirect</payment_page>
                 <one_click>0</one_click>
                 <pendingpayment_order_status>pending_payment</pendingpayment_order_status>
                 <processing_order_status>processing</processing_order_status>
                 <canceled_order_status>canceled</canceled_order_status>
                 <allowspecific>0</allowspecific>
                 <generate_invoice>1</generate_invoice>
+                <sort_order>0</sort_order>
+                <model>Payplug\Payments\Model\Payment\Standard</model>
+                <payment_action>authorize_capture</payment_action>
+            </payplug_payments_standard>
+        </payment>
+        <payplug_payments>
+            <general>
+                <environmentmode>test</environmentmode>
+                <payment_page>redirect</payment_page>
                 <connected>0</connected>
                 <verified>0</verified>
-                <premium>0</premium>
-                <sort_order>0</sort_order>
-                <model>Payplug\Payments\Model\PaymentMethod</model>
-                <payment_action>authorize_capture</payment_action>
-            </payplug_payments>
-        </payment>
+            </general>
+        </payplug_payments>
     </default>
 </config>
